Adjust mvp_pattern in mainservice3
